refactor: extract props, emits from default.ts to [component-name].ts
