+++ packages/components/select-v2/src/group-item.ts
@@ -1,0 +1,8 @@
+export const GroupProps = {
+  item: {
+    type: Object,
+    required: true,
+  },
+  style: Object,
+  height: Number,
+}
