+++ packages/components/select-v2/src/option-item.ts
@@ -1,0 +1,12 @@
+export const OptionProps = {
+  data: Array,
+  disabled: Boolean,
+  hovering: Boolean,
+  item: Object,
+  index: Number,
+  style: Object,
+  selected: Boolean,
+  created: Boolean,
+}
+
+export const OptionsEmits = ['select', 'hover']
